[models] Add possibility to validate inverse relations #4
Added a reverse validation method which ensures that before the
organization of an object is changed, no object is related to it, else a
validation error is raised.
The method receives as arguments the models which are related to the
current entity and count for that the possible relations to this entity
with this particular organization.

Fixes #4