[admin] Fixed show the org related to that user #44
Solved the problem by removing the is_admin=True from the operator_org
queryset and overriding the has_permission method of the
OrganizationAdmin class to return false if an operator requesting to
change an organization is not an admin of that organization.
Created a test case in which an operator is created with is_staff set to
true and user permissions added to him. 2 organizations were created:
org1, org2 and operator was made member of these orgs and admin of
org1. the change page of the operator was visited and checked if
operator could see all the orgs which he is a member of or not. Equally,
it tested if the operator could change org1 and org2. test passes for
org1 but fails for org2 since operator is not an admin in org2.

Fixes #44