[change/fix] Multi-tenant admin classes now allow only org managers #164
Multi-tenant admin classes now allow only org managers.

Before this version, a user needed to be only org member
to see items of that organization in the admin, but this
is wrong! An ``OrganizationUser`` which has ``is_admin=False`` is
only an end-user of that organization.
Instead, an ``OrganizationUser`` which has ``is_admin=True`` is
also a manager and only this type of user shall be allowed
to manage items of the organization through the django admin site.

This is needed in order to support users being simple end-users
in one organization but administrators in others, otherwise
a staff user who is administrator of one organization would be
able to change also items of other organizations where
they are only members and not managers.

Fixes #164