[chores] Allow passing a string or uuid to the membership helpers
Allow passing an UUID or str to the membership helpers, which will
be interpreted as organization primary key.
This is useful in those cases in which building the organization
instance requires generating an extra DB query: passing
obj.organization_id instead of obj.organization will result
in less queries being executed.